refactor: move api calls of Assembly page to api.ts
