+++ src/modules/Dashboard/modules/pages/Assembly/AssemblyTable.tsx
@@ -1,11 +1,9 @@
 import React, { Dispatch, FC, SetStateAction, useEffect, useState } from 'react'
 import { initialState, selectProps } from '../../utils/setupUtils'
-import { setupRoutes, tableRoutes } from '../../../../../services/urls'
-import { privateGateway } from '../../../../../services/apiGateway'
 import Modal from './AssemblyModal'
 import { CustomSelect } from '../../../components/CustomSelect/CustomSelect'
 import CustomTable from '../../components/CustomTable/CustomTable'
-import * as yup from 'yup'
+import { fetchAssemblys, fetchDistricts } from './assemblyAPI'
 
 interface AssemblySetupProps {
     setViewSetup: Dispatch<SetStateAction<boolean>>
@@ -141,25 +139,7 @@ function rawString(str: string) {
     str = str.replaceAll(' ', '')
     return str
 }
-function fetchDistricts(setData: Dispatch<SetStateAction<selectProps[]>>) {
-    privateGateway.get(setupRoutes.district.list)
-        .then(res => res.data.response.districts)
-        .then(data => setData(data))
-        .catch(err => console.log('Error :', err?.response?.data?.message?.general[0]))
-}
-function fetchAssemblys(
-    setData: Dispatch<SetStateAction<AssemblyTableProps[]>>,
-    setData2: Dispatch<SetStateAction<AssemblyTableProps[]>>,
-    updateTable?: Function
-) {
-    privateGateway.get(tableRoutes.assembly.list)
-        .then(res => res.data.response)
-        .then(data => {
-            setData(data)
-            setData2(data)
-            if (updateTable) updateTable(data)
-        })
-        .catch(err => console.log('Error :', err?.response?.data?.message?.general[0]))
-}
+
+
 export default AssemblyTable
